+++ ceammc/CHANGELOG.md
@@ -35,6 +35,7 @@
   - [choose( method added to get random element from \[data.set\], \[local.set\] and \[global.set\]
   - [choose( method added to get random element from \[data.dict\], \[local.dict\] and \[global.dict\]
   - [pause( method added to \[array.grainer~\]
+  - [slice( method added to \[array.grainer~\] - to slice source to specified number of equal grains
 - misc:
   - conv.lin2exp: list support added
   - conv.lin2curve: list support added